requirements.txt
PyQt5>=5.15.0
pyyaml>=5.1.0
markdown>=3.3.0
pytest>=6.0.0
pytest-cov>=2.10.0
pyinstaller>=5.6.0
